remove old tipSelection nad rename tipManager to tipSolidifier
